+++ include/play.h
@@ -3,7 +3,7 @@
 
 void Play_init(char* path);
 void Play_change_scroll_speed(int diff);
-void Play_update(double dt);
+void Play_update(long dt);
 void Play_draw();
 
 #endif
